feat: support test module
Only the test of mov instruction is implement so far.

The test logic should be optimized later. The current logic is just calling the
generate function with an operand type every time. It can just test if operand
type is implemented, but can not make sure if it's implemented correctly.

It's better to create test files using a script. As all the instruction and all
the relevant operand type is written in insnsa.c.